- move the useful functions in `/src/html_deps_parser/funcdefs.ts` to `/src/funcdefs.ts`, and refactor the imports. - purge `/src/html_deps_parser/funcdefs.ts`. - move the remaining three parser files (and type declarations) in `/src/html_deps_parser/*.ts` to `/extra/html_deps_parser/*.ts`, and refactor. - ready to merge the branch `cleanup-1` to `main`.
